Add display prop to Text
The styles associated with it are set in global.css,
because they would typically be part of a theme, are complex
enough not to be practical to override with CSS variables,
and will likely be re-used widely across different components.

On Apple devices, system fonts are preferred for performance.
On the other devices, Inter is used because it will offer a very
similar visual experience. (Thanks @marchamm for the tip!)

See https://seek-oss.github.io/capsize
and https://polaris.shopify.com/design/typography#section-font-stack